+++ apps/app/src/app/app.module.ts
@@ -1,20 +1,62 @@
 import { APP_BASE_HREF } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
 import { RouteReuseStrategy } from '@angular/router';
+import {
+  ClrIconModule,
+  ClrLayoutModule,
+  ClrVerticalNavModule,
+} from '@clr/angular';
 import { environment } from '../environments/environment';
 import { AppRoutingModule } from './app-routing.module';
 import { AppTranslateModule } from './app-translate.module';
 import { AppComponent } from './app.component';
-import { LanguageService, CoreModule, RemoveRouteReuseStrategy } from './pages';
+import {
+  TheApplicationsSubnavComponent,
+  TheFavoritesModalComponent,
+  TheFavoritesNavigationComponent,
+  TheHeaderTimeComponent,
+  TheHistoryNavigationComponent,
+  TheLanguagesSelectorComponent,
+} from './shared/feature';
+import {
+  CoreModule,
+  LanguageService,
+  RemoveRouteReuseStrategy,
+} from './shared/util';
+
+import '@clr/icons';
+import '@clr/icons/shapes/chart-shapes';
+import '@clr/icons/shapes/essential-shapes';
+import '@clr/icons/shapes/social-shapes';
+import '@clr/icons/shapes/technology-shapes';
+import { TheHeaderComponent } from './shared/ui';
 
 /**
  * Load app data
  */
 @NgModule({
-  imports: [CoreModule, AppRoutingModule, AppTranslateModule],
+  imports: [
+    ClrLayoutModule,
+    ClrVerticalNavModule,
+    ClrIconModule,
+    CoreModule,
+    AppRoutingModule,
+    AppTranslateModule,
+    BrowserModule,
+    TheApplicationsSubnavComponent,
+    TheLanguagesSelectorComponent,
+    TheHeaderComponent,
+    TheHistoryNavigationComponent,
+    TheFavoritesNavigationComponent,
+    TheFavoritesModalComponent,
+    TheHeaderTimeComponent,
+  ],
+
   providers: [
     { provide: APP_BASE_HREF, useValue: environment.baseHref },
     { provide: RouteReuseStrategy, useClass: RemoveRouteReuseStrategy },
+    { provide: Window, useValue: window },
   ],
   declarations: [AppComponent],
   bootstrap: [AppComponent],
