+++ apps/app/src/app/error/feature/error-detail/error-detail-routing.module.ts
