--- apps/app/src/app/pages/index.ts
@@ -1,24 +1,0 @@
-export {
-  TheApplicationsSubnavComponent,
-  TheFavoritesModalComponent,
-  TheFavoritesNavigationComponent,
-  TheHeaderTimeComponent,
-  TheHistoryNavigationComponent,
-  TheLanguagesSelectorComponent,
-} from './feature';
-
-export { TheHeaderComponent } from './ui';
-
-export {
-  AppNavLink,
-  BrowserTitleService,
-  CoreModule,
-  ExternalServicesEnum,
-  GrafanaService,
-  HistoryService,
-  LanguageCode,
-  LanguageService,
-  RemoveRouteReuseStrategy,
-  SeqService,
-  SettingsService,
-} from './util';
