--- apps/app/src/app/pages/pages.component.html
@@ -1,74 +1,0 @@
-<clr-main-container>
-  <clr-header>
-    <div class="header-nav">
-      <div class="branding">
-        <a routerLink="/">
-          <span class="title">
-            <figure>
-              <img src="assets/images/armonik-logo.png" alt="ArmoniK logo" />
-            </figure>
-            &nbsp; - &nbsp;
-            <app-pages-the-header-time></app-pages-the-header-time>
-          </span>
-        </a>
-      </div>
-    </div>
-    <div class="header-actions">
-      <app-pages-the-languages-selector></app-pages-the-languages-selector>
-    </div>
-  </clr-header>
-  <div class="subnav no-scrollbar">
-    <app-pages-the-applications-subnav></app-pages-the-applications-subnav>
-  </div>
-  <div class="content-container">
-    <clr-vertical-nav [clr-nav-level]="1">
-      <a
-        *ngFor="let link of links; trackBy: trackByLabel"
-        clrVerticalNavLink
-        [routerLink]="link.path"
-        [queryParams]="link.queryParams"
-        routerLinkActive="active"
-      >
-        <clr-icon [attr.shape]="link.shape" clrVerticalNavIcon></clr-icon>
-        {{ link.label | translate }}</a
-      >
-      <ng-container
-        *ngIf="{
-          seq: settingsService.seqSubject$ | async,
-          grafana: settingsService.grafanaSubject$ | async
-        } as isUp"
-      >
-        <div class="nav-divider" *ngIf="isUp.seq || isUp.grafana"></div>
-        <a clrVerticalNavLink href="/seq/" target="_blank" *ngIf="isUp.seq">
-          <clr-icon shape="crosshairs" clrVerticalNavIcon></clr-icon>
-          Seq</a
-        >
-        <a
-          clrVerticalNavLink
-          href="/grafana/"
-          target="_blank"
-          *ngIf="isUp.grafana"
-        >
-          <clr-icon shape="line-chart" clrVerticalNavIcon></clr-icon>
-          Grafana</a
-        >
-      </ng-container>
-      <div class="nav-divider"></div>
-      <app-pages-the-favorites-navigation></app-pages-the-favorites-navigation>
-      <app-pages-the-history-navigation></app-pages-the-history-navigation>
-      <div class="nav-divider"></div>
-      <a clrVerticalNavLink routerLink="/how-to-use" routerLinkActive="active">
-        <clr-icon shape="help" clrVerticalNavIcon></clr-icon>
-        {{ 'navigation.how_to_use' | translate }}
-      </a>
-      <div class="version">
-        <span> v0.7.0 </span>
-      </div>
-    </clr-vertical-nav>
-
-    <div class="content-area">
-      <app-pages-the-favorites-modal> </app-pages-the-favorites-modal>
-      <router-outlet></router-outlet>
-    </div>
-  </div>
-</clr-main-container>
