--- apps/app/src/app/pages/pages.component.scss
@@ -1,55 +1,0 @@
-[actions] .btn {
-  min-width: 0;
-}
-
-.title {
-  display: flex;
-  flex-direction: row;
-  justify-items: center;
-  align-items: center;
-
-  & figure {
-    margin: 0;
-    & img {
-      height: 1rem;
-      width: auto;
-    }
-  }
-}
-
-app-pages-the-languages-selector {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-app-pages-the-applications-subnav {
-  width: 100%;
-}
-
-.version {
-  flex-grow: 1;
-
-  display: flex;
-  flex-direction: row;
-  align-items: flex-end;
-  justify-content: center;
-
-  margin-bottom: 0.9rem;
-}
-
-.content-area {
-  position: relative;
-}
-
-app-pages-the-favorites-modal {
-  position: absolute;
-  top: 1rem;
-  right: 0.5rem;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-  align-items: center;
-}
