+++ apps/app/src/app/results/feature/results-list/results-list.page.scss
