+++ apps/app/src/app/results/feature/results-list/results-list.page.ts
@@ -1,21 +1,21 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { GrpcSessionsService } from '@armonik.admin.gui/sessions/data-access';
-import { DisabledIntervalValue } from '@armonik.admin.gui/shared/feature';
+import { GrpcResultsService } from '@armonik.admin.gui/results/data-access';
 import {
-  GetSessionResponse,
   GrpcPagerService,
-  ListSessionsRequest,
-  ListSessionsResponse,
-  SessionStatus,
+  ListResultsRequest,
+  ListResultsResponse,
+  ResultRaw,
+  ResultStatus,
 } from '@armonik.admin.gui/shared/data-access';
+import { DisabledIntervalValue } from '@armonik.admin.gui/shared/feature';
 import { ClrDatagridSortOrder, ClrDatagridStateInterface } from '@clr/angular';
 import {
   BehaviorSubject,
   catchError,
   concatMap,
   distinctUntilChanged,
-  first,
+  interval,
   map,
   merge,
   Observable,
@@ -26,36 +26,28 @@ import {
   tap,
   timer,
 } from 'rxjs';
-import { BrowserTitleService, LanguageService } from '../../../util';
+import {
+  BrowserTitleService,
+  LanguageService,
+  SettingsService,
+} from '../../../shared/util';
 
 @Component({
-  selector: 'app-pages-sessions-list',
-  templateUrl: './sessions-list.page.html',
-  styleUrls: ['./sessions-list.page.scss'],
+  selector: 'app-pages-results-list',
+  templateUrl: './results-list.page.html',
+  styleUrls: ['./results-list.page.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class SessionsListComponent implements OnInit {
+export class ResultsListComponent implements OnInit {
   private _state: ClrDatagridStateInterface = {};
 
-  /** Get a single session */
-  private _opened$ = new BehaviorSubject<boolean>(false);
-  private _subjectSingleSession = new Subject<string>();
-  private _triggerSingleSession = this._subjectSingleSession.asObservable();
-
-  loadingSingleSession$ = new BehaviorSubject<string | null>(null);
-  loadSingleSession$ = this._triggerSingleSession.pipe(
-    tap((sessionId) => this.loadingSingleSession$.next(sessionId)),
-    switchMap((sessionId) => this._getSession$(sessionId))
-  );
-
-  /** Get sessions */
   private _subjectManual = new Subject<void>();
   private _subjectDatagrid = new Subject<ClrDatagridStateInterface>();
   private _intervalValue = new Subject<number>();
   private _stopInterval = new Subject<void>();
   public stopInterval$ = this._stopInterval.asObservable();
 
-  /** Triggers to reload sessions */
+  /** Triggers to reload data */
   private _triggerManual$ = this._subjectManual.asObservable();
   private _triggerDatagrid$ = this._subjectDatagrid.asObservable().pipe(
     tap((state) => this._saveState(state)),
@@ -75,16 +67,16 @@ export class SessionsListComponent implements OnInit {
       switchMap((time) => timer(0, time).pipe(takeUntil(this.stopInterval$)))
     );
 
-  loadingSessions$ = new BehaviorSubject<boolean>(true);
-  totalSessions$ = new BehaviorSubject<number>(0);
+  loadingResults$ = new BehaviorSubject<boolean>(true);
+  totalResults$ = new BehaviorSubject<number>(0);
 
-  loadSessions$ = merge(
+  loadResults$ = merge(
     this._triggerManual$,
     this._triggerDatagrid$,
     this._triggerInterval$
   ).pipe(
-    tap(() => this.loadingSessions$.next(true)),
-    switchMap(() => this._listSessions$())
+    tap(() => this.loadingResults$.next(true)),
+    switchMap(() => this._listResults$())
   );
 
   constructor(
@@ -92,22 +84,31 @@ export class SessionsListComponent implements OnInit {
     private _activatedRoute: ActivatedRoute,
     private _browserTitleService: BrowserTitleService,
     private _languageService: LanguageService,
-    private _grpcSessionsService: GrpcSessionsService,
+    private _settingsService: SettingsService,
+    private _grpcResultsService: GrpcResultsService,
     private _grpcPagerService: GrpcPagerService
   ) {}
 
   ngOnInit(): void {
     this._browserTitleService.setTitle(
-      this._languageService.instant('pages.sessions-list.title')
+      this._languageService.instant('results.title')
     );
   }
 
   public get OrderByField() {
-    return ListSessionsRequest.OrderByField;
+    return ListResultsRequest.OrderByField;
   }
 
-  public get SessionStatusEnum() {
-    return SessionStatus;
+  public get ResultsStatusEnum() {
+    return ResultStatus;
+  }
+
+  public get intervals() {
+    return this._settingsService.intervals;
+  }
+
+  public get initialInterval() {
+    return this._settingsService.initialInterval;
   }
 
   public onUpdateInterval(value: number) {
@@ -120,7 +121,7 @@ export class SessionsListComponent implements OnInit {
   }
 
   public defaultSortOrder(
-    field: ListSessionsRequest.OrderByField
+    field: ListResultsRequest.OrderByField
   ): ClrDatagridSortOrder {
     const orderBy = Number(
       this._activatedRoute.snapshot.queryParamMap.get('orderBy')
@@ -137,70 +138,34 @@ export class SessionsListComponent implements OnInit {
   }
 
   /**
-   * Cancel a session
+   * Track by interval
    *
-   * @param sessionId
-   */
-  public cancelSession(sessionId: string): void {
-    this._grpcSessionsService
-      .cancel$(sessionId)
-      .pipe(
-        first(),
-        catchError((error: Error) => {
-          console.error(error);
-
-          return of({} as ListSessionsResponse);
-        })
-      )
-      .subscribe({
-        next: () => this.manualRefreshSessions(),
-      });
-  }
-
-  /**
-   * Call a new session
+   * @param _
+   * @param interval
    *
-   * @param sessionId
+   * @returns Interval
    */
-  public viewSessionDetail(sessionId: string): void {
-    this._subjectSingleSession.next(sessionId);
+  public trackByInterval(_: number, interval: number): string {
+    return interval.toString();
   }
 
   /**
-   * Open modal to view details
-   */
-  public openGetSessionModal(): void {
-    this._opened$.next(true);
-  }
-
-  /**
-   * Close modal to view details
-   */
-  public closeGetSessionModal(): void {
-    this._opened$.next(false);
-  }
-
-  public get isGetSessionModalOpened$(): Observable<boolean> {
-    return this._opened$.asObservable();
-  }
-
-  /**
-   * Refresh sessions using a new state
+   * Refresh results using a new state
    *
    * @param state
    *
    * @returns void
    */
-  public refreshSessions(state: ClrDatagridStateInterface): void {
+  public refreshResults(state: ClrDatagridStateInterface): void {
     this._subjectDatagrid.next(state);
   }
 
   /**
-   * Refresh manually sessions using new state
+   * Refresh manually results using new state
    *
    * @returns void
    */
-  public manualRefreshSessions(): void {
+  public manualRefreshResults(): void {
     this._subjectManual.next();
   }
 
@@ -217,6 +182,18 @@ export class SessionsListComponent implements OnInit {
       map((value) => Number(value)),
       distinctUntilChanged()
     );
+  }
+
+  /**
+   * Track by result
+   *
+   * @param _
+   * @param result
+   *
+   * @returns Id
+   */
+  public trackByResult(_: number, result: ResultRaw): string {
+    return result.name ?? '';
   }
 
   /**
@@ -238,42 +215,23 @@ export class SessionsListComponent implements OnInit {
   }
 
   /**
-   * List sessions
+   * List results
    *
-   * @returns Observable<ListSessionsResponse>
+   * @returns Observable<ListResultsResponse>
    */
-  private _listSessions$(): Observable<ListSessionsResponse> {
+  private _listResults$(): Observable<ListResultsResponse> {
     const params = this._grpcPagerService.createParams(this._restoreState());
 
-    return this._grpcSessionsService.list$(params).pipe(
-      catchError((error: Error) => {
+    return this._grpcResultsService.list$(params).pipe(
+      catchError((error) => {
         console.error(error);
         this._stopInterval.next();
 
-        return of({} as ListSessionsResponse);
+        return of({} as ListResultsResponse);
       }),
-      tap((sessions) => {
-        this.loadingSessions$.next(false);
-        this.totalSessions$.next(sessions.total ?? 0);
-      })
-    );
-  }
-
-  /**
-   * Get single session
-   *
-   * @returns Observable<GetSessionResponse>
-   */
-  private _getSession$(sessionId: string): Observable<GetSessionResponse> {
-    return this._grpcSessionsService.get$(sessionId).pipe(
-      catchError((error: Error) => {
-        console.error(error);
-
-        return of({} as GetSessionResponse);
-      }),
-      tap(() => {
-        this.openGetSessionModal();
-        this.loadingSingleSession$.next(null);
+      tap((results) => {
+        this.loadingResults$.next(false);
+        this.totalResults$.next(results.total ?? 0);
       })
     );
   }
