+++ apps/app/src/app/sessions/feature/sessions-list/sessions-list.page.scss
